requirements.txt
beautifulsoup4==4.13.5
Flask==3.1.2
Pillow==11.3.0
Requests==2.32.5
tinydb==4.8.2
